requirements.txt
numpy
pillow
fastapi